requirements.txt
asgiref==3.8.0
click==8.1.7
Django==4.2.10
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.8.0
pytz==2023.3
services==0.1.1
six==1.16.0
sqlparse==0.4.4
